+++ src/commands/deprecated/seedpack.ts
@@ -1,0 +1,9 @@
+import { KlasaMessage } from 'klasa';
+
+import { BotCommand } from '../../lib/structures/BotCommand';
+
+export default class extends BotCommand {
+	async run(msg: KlasaMessage) {
+		return msg.channel.send('You now open seedpacks using `/open name:Seed pack`');
+	}
+}
